requirements.txt

numpy
PyQt5
pyqtgraph
scipy
jupyter
py-ecg-detectors